backend/pyproject.toml
[tool.poetry]
name = "razmenakarata"
version = "0.1.0"
description = ""
authors = ["Aleksa Ognjanovic <alexa.ognjanovic@protonmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
djangorestframework = "^3.15.1"
djoser = "^2.2.2"
python-dotenv = "^1.0.1"
redis = "^5.0.4"
psycopg2-binary = "^2.9.9"
django-filter = "^24.2"
markdown = "^3.6"
lxml = "^5.2.1"
requests = "^2.31.0"
django-cors-headers = "^4.3.1"
drf-nested-routers = "^0.93.5"
sorl-thumbnail = "^12.10.0"
pillow = "^10.3.0"
drf-spectacular = "^0.27.2"
drf-standardized-errors = "^0.13.0"
celery = {extras = ["redis"], version = "^5.4.0"}
django-celery-beat = "^2.6.0"
channels = {extras = ["daphne"], version = "^4.1.0"}
channels-redis = "^4.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
